Implement Core Game Systems pt2
-GameManager refactoring: now acts as an orchestrator
--Minor Tweaks & Cleanup
       Adjusted Card flip/unflip animations ( Do Tween)
-- Added more cards

+++ Assets/GameFolder/_Scripts/Save System/SaveData.cs
@@ -7,7 +7,7 @@ namespace SKC.SaveSystem
     public class SaveData
     {
         public int level;
-        public int highestPoint;
+        public float highestPoint;
         public int totalMatchesFound;  
         public int totalMovesMade;     
         public float gameVolume;
